Remove v1 deployment parameters code
