+++ src/components/FlowRunMenu.vue
@@ -21,12 +21,7 @@
     :flow-run="flowRun"
   />
 
-  <template v-if="can.access.schemasV2">
-    <FlowRunResumeModalV2 v-model:showModal="showResumeModal" :flow-run-id="flowRunId" />
-  </template>
-  <template v-else>
-    <FlowRunResumeModal v-model:showModal="showResumeModal" :flow-run-id="flowRunId" />
-  </template>
+  <FlowRunResumeModal v-model:showModal="showResumeModal" :flow-run-id="flowRunId" />
 
   <FlowRunCancelModal
     v-model:showModal="showCancelModal"
@@ -57,8 +52,8 @@
 <script lang="ts" setup>
   import { showToast } from '@prefecthq/prefect-design'
   import { computed, ref, toRefs } from 'vue'
-  import { FlowRunRetryModal, FlowRunResumeModal, FlowRunCancelModal, FlowRunSuspendModal, ConfirmStateChangeModal, ConfirmDeleteModal, CopyOverflowMenuItem } from '@/components'
-  import FlowRunResumeModalV2 from '@/components/FlowRunResumeModalV2.vue'
+  import { FlowRunRetryModal, FlowRunCancelModal, FlowRunSuspendModal, ConfirmStateChangeModal, ConfirmDeleteModal, CopyOverflowMenuItem } from '@/components'
+  import FlowRunResumeModal from '@/components/FlowRunResumeModal.vue'
   import { useCan, useWorkspaceApi, useShowModal, useWorkspaceRoutes, useFlowRuns, useFlowRun, useDeployment } from '@/compositions'
   import { localization } from '@/localization'
   import { FlowRunsFilter, isPausedStateType, isRunningStateType, isStuckStateType, isTerminalStateType, StateUpdateDetails } from '@/models'
