+++ src/components/QuickRunParametersModal.vue
@@ -1,12 +1,20 @@
 <template>
-  <p-modal v-model:showModal="internalShowModal" class="parameters-modal" title="Run Deployment">
-    <p-form @submit="submit">
-      <SchemaInput v-model="parameters" :schema="deployment.parameterOpenApiSchema" />
-    </p-form>
+  <p-modal v-model:showModal="internalShowModal" class="quick-run-parameters-modal-v2" title="Run Deployment">
+    <p-content>
+      <div class="quick-run-parameters-modal-v2__header">
+        <h3>{{ localization.info.parameters }}</h3>
+        <p-icon-button-menu small>
+          <p-overflow-menu-item v-if="kind === 'json'" label="Use form input" @click="kind = 'none'" />
+          <p-overflow-menu-item v-if="kind === 'none'" label="Use JSON input" @click="kind = 'json'" />
+        </p-icon-button-menu>
+      </div>
+
+      <SchemaFormV2 :id="formId" v-model:values="parameters" :schema="deployment.parameterOpenApiSchemaV2" :kinds="['json']" @submit="submit" />
+    </p-content>
 
     <template #actions>
       <slot name="actions">
-        <p-button primary @click="submit">
+        <p-button type="submit" primary :form="formId">
           Run
         </p-button>
       </slot>
@@ -15,17 +23,15 @@
 </template>
 
 <script lang="ts" setup>
-  import { PButton, showToast } from '@prefecthq/prefect-design'
-  import { useValidationObserver } from '@prefecthq/vue-compositions'
-  import { useField } from 'vee-validate'
-  import { computed, h } from 'vue'
+  import { PButton, randomId, showToast } from '@prefecthq/prefect-design'
+  import { computed, h, ref } from 'vue'
   import { useRouter } from 'vue-router'
-  import { ToastFlowRunCreate, SchemaInput } from '@/components'
+  import { ToastFlowRunCreate } from '@/components'
   import { useWorkspaceApi, useWorkspaceRoutes } from '@/compositions'
-  import { useForm } from '@/compositions/useForm'
   import { localization } from '@/localization'
-  import { Deployment, DeploymentFlowRunCreate } from '@/models'
-  import { SchemaValues } from '@/types/schemas'
+  import { Deployment, DeploymentFlowRunCreateV2 } from '@/models'
+  import { SchemaFormV2, SchemaValuesV2 } from '@/schemas'
+  import { usePrefectKind } from '@/schemas/compositions/usePrefectKind'
   import { getApiErrorMessage } from '@/utilities/errors'
 
   const props = defineProps<{
@@ -36,17 +42,14 @@
   const api = useWorkspaceApi()
   const router = useRouter()
   const routes = useWorkspaceRoutes()
+  const formId = randomId()
 
   const emit = defineEmits<{
     (event: 'update:showModal', value: boolean): void,
   }>()
 
-  const { handleSubmit } = useForm<DeploymentFlowRunCreate>({
-    initialValues: {
-      parameters: props.deployment.parameters,
-      schema: props.deployment.parameterOpenApiSchema,
-    },
-  })
+  const parameters = ref<SchemaValuesV2>({ ...props.deployment.parametersV2 })
+  const { kind } = usePrefectKind(parameters)
 
   const internalShowModal = computed({
     get() {
@@ -57,31 +60,17 @@
     },
   })
 
-  const { validate } = useValidationObserver()
-
-  const submit = handleSubmit(async (values): Promise<void> => {
-    const valid = await validate()
-
-    if (!valid) {
-      return
-    }
-
-    const resolvedValues: DeploymentFlowRunCreate = {
+  async function submit(): Promise<void> {
+    const values: DeploymentFlowRunCreateV2 = {
       state: {
         type: 'scheduled',
         message: 'Run from the Prefect UI',
       },
-      ...values,
+      parameters: parameters.value,
     }
-    await createDeploymentFlowRun(props.deployment.id, resolvedValues)
-    internalShowModal.value = false
-  })
 
-  const { value: parameters } = useField<SchemaValues>('parameters')
-
-  async function createDeploymentFlowRun(deploymentId: string, value: DeploymentFlowRunCreate): Promise<void> {
     try {
-      const flowRun = await api.deployments.createDeploymentFlowRun(deploymentId, value)
+      const flowRun = await api.deployments.createDeploymentFlowRunV2(props.deployment.id, values)
 
       const toastMessage = h(ToastFlowRunCreate, { flowRun, flowRunRoute: routes.flowRun, router, immediate: true })
       showToast(toastMessage, 'success')
@@ -90,5 +79,15 @@
       showToast(message, 'error')
       console.error(error)
     }
+
+    internalShowModal.value = false
   }
 </script>
+
+<style>
+.quick-run-parameters-modal-v2__header { @apply
+  flex
+  items-center
+  justify-between
+}
+</style>
