+++ src/services/can.ts
@@ -4,7 +4,6 @@ import { MaybeRef } from '@/types/reactivity'
 export const workspaceFeatureFlags = [
   'access:deploymentStatus',
   'access:workQueueStatus',
-  'access:schemasV2',
   'access:flowRunInfraOverrides',
 ] as const satisfies Readonly<`access:${string}`[]>
